refactor(evidence): split mandate core types facade
